requirements.txt
pytest-cov
fastapi 
sqlalchemy
sqlalchemy_utils
pydantic 
alembic 
psycopg2
asyncpg
python-dotenv